Add Drehem Network Graphs
These graph files are the preliminary results of network analysis used as a prosopographical disambiguation method for the ca. 15,000 Sumerian Ur III texts (ca. 2100-2004 bce), which have been attributed to the province of Drehem, ancient Puzriš-Dagan. These network visualize the results from the first iteration (Drehem_01) of the merged network, a dynamic network, in two file formats. This also includes the eighth+ iteration (Drehem_08t2) of the merged network, a static network, with a time interval.